Messed around with exponent stuff.  Put the login in ExponentiationEvaluator where it should be and have it throwing an ArithmeticException that we can catch.

+++ Calculator/src/main/java/edu/odu/cs/cs600/calculator/math/MathUtil.java
@@ -47,7 +47,7 @@ public class MathUtil
 		return subtractionEvaluator.compute(x, y);
 	}
 	
-	public static double exponentiate(double x, int y) throws ArithmeticException {
+	public static double exponentiate(double x, double y) throws ArithmeticException {
 		return exponentiationEvaluator.compute(x, y);
 	}
